+++ packages/safe-ds-lang/src/language/generation/python/constants.ts
@@ -1,0 +1,1 @@
+export const CODEGEN_PREFIX = '__gen_';
